pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-product-data"
version = "0.1.0"
description = "Open-source AI startup data repository with automated scrapers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "jsonschema>=4.21",
    "python-dateutil>=2.9",
    "python-Levenshtein>=0.25",
    "tenacity>=8.2",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.39",
]
firecrawl = [
    "firecrawl-py>=1.0",
]
appstores = [
    "google-play-scraper>=1.2",
]
browser = [
    "playwright>=1.40",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "playwright>=1.40",
]

[project.scripts]
aiscrape = "scrapers.cli:cli"

[tool.setuptools.packages.find]
include = ["scrapers*", "scrapers.data*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]
